requirements.txt
markdownify
smolagents==1.13.0
requests
ddgs
pandas
gradio==5.23.1
transformers
huggingface_hub
python-dotenv
pytz
pyyaml
openai>=1.0.0
